refactored testing and separated out nowcast model into separate file
