+++ test/original_seir_model_test.py
@@ -10,12 +10,8 @@ from matplotlib import pyplot as plt
 from datetime import datetime, timedelta
 
 from pyseir.models.seir_model import (
-    Demographics,
-    ramp_function,
     SEIRModel,
     steady_state_ratios,
-    NowcastingSEIRModel,
-    ModelRun,
 )
 from pyseir.rt.constants import InferRtConstants
